Add TerminalApp class to implement MVC pattern
Previously to this commit WikiRider was communicating directly with
RidePrinter. This can't happen if we expect to handle several ways of UI,
one solution is to isolate WikiRider implementation from other classes.
To accomplish this, I've created TerminalApp to control
and coordinate both WikiRider and RidePrinter.

With this modification WikiRider only cares about the logic (model),
RidePrinter about the presentation (view) and TerminalApp about
controlling the whole application (controler).

It's important to mention that now WikiRider.run() returns a generator
to handle each webpage it visits.

This MVC architecture provides the basic structure
to provide a GUI in the future.